feat: add formatted methods (#4)
This change adds the following methods for formatting log messages:

- `Printf`
- `Debugf`
- `Infof`
- `Warnf`
- `Errorf`

Arguments are handled in the manner of `log.Printf`: the string is
formatted according to the specifier + provided arguments, then returned
with a trailing newline.

Use of `Println`/etc. is still preferred, as `clogs` merely bails out to
`fmt.Sprintf` for formatting. It is still slightly more performant than
the standard logger, but it commonly requires more allocations per log
operation.